requirements.txt
pandas
numpy
certifi
python-dotenv
pymongo

-e .